feat(MainPanel): Add MainPanel
Created MainPanel component which is a generic component and can be used in WeeklyForecastDetailsModal and TodaysWeather. It has two props: tempCelsius and weatherCondition, and based on these, it will display the weather and an icon for specific day.
